feat: patch for extensions

+++ tools/build_extension.sh
@@ -4,12 +4,20 @@ set -eux -o pipefail
 
 cd "$(dirname "$0")/.."
 
+if ! git --no-pager diff --exit-code || ! git --no-pager diff --cached --exit-code; then
+  echo "changes are detected in working copy."
+  exit 1
+fi
+
 which blender
 blender --version
 
 pushd src/io_scene_vrm
 git clean -xdff .
 popd
+
+patch -p1 <extension.patch
+git --no-pager diff
 
 rm -fr extension_output
 mkdir extension_output
@@ -20,3 +28,5 @@ blender --command extension build \
 
 blender --command extension validate \
   extension_output/vrm_extension.zip
+
+git checkout .
